requirements.txt
requests>=2.32.0
openai>=2.15.0
python-dotenv>=1.0.0
